pyproject.toml
[tool.poetry]
name = "token-store"
version = "0.1.0"
description = ""
authors = ["canaparro <marcelo.canaparro@pagefreezer.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
tortoise-orm = {extras = ["aiosqlite"], version = "^0.21.4"}
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
